+++ src/executor/CMakeLists.txt
@@ -28,8 +28,8 @@ nebula_add_library(
     query/DataCollectExecutor.cpp
     query/DataJoinExecutor.cpp
     query/IndexScanExecutor.cpp
-    query/ConjunctPathExecutor.cpp
-    query/BFSShortestPathExecutor.cpp
+    algo/ConjunctPathExecutor.cpp
+    algo/BFSShortestPathExecutor.cpp
     admin/SwitchSpaceExecutor.cpp
     admin/CreateUserExecutor.cpp
     admin/DropUserExecutor.cpp
