+++ src/planner/Algo.cpp
@@ -1,0 +1,13 @@
+/* Copyright (c) 2020 vesoft inc. All rights reserved.
+ *
+ * This source code is licensed under Apache 2.0 License,
+ * attached with Common Clause Condition 1.0, found in the LICENSES directory.
+ */
+
+#include "planner/Algo.h"
+
+namespace nebula {
+namespace graph {
+
+}  // namnspace graph
+}  // namespace nebula
