+++ src/planner/Query.h
@@ -947,32 +947,6 @@ private:
     std::vector<Expression*>                hashKeys_;
     std::vector<Expression*>                probeKeys_;
 };
-
-class ProduceSemiShortestPath : public PlanNode {
-};
-
-class BFSShortestPath : public SingleInputNode {
-public:
-    static BFSShortestPath* make(QueryContext* qctx, PlanNode* input) {
-        return qctx->objPool()->add(new BFSShortestPath(qctx->genId(), input));
-    }
-
-private:
-    BFSShortestPath(int64_t id, PlanNode* input)
-        : SingleInputNode(id, Kind::kBFSShortest, input) {}
-};
-
-class ConjunctPath : public BiInputNode {
-public:
-    static ConjunctPath* make(QueryContext* qctx, PlanNode* left, PlanNode* right) {
-        return qctx->objPool()->add(new ConjunctPath(qctx->genId(), left, right));
-    }
-
-private:
-    ConjunctPath(int64_t id, PlanNode* left, PlanNode* right)
-        : BiInputNode(id, Kind::kConjunctPath, left, right) {}
-};
-
 }  // namespace graph
 }  // namespace nebula
 #endif  // PLANNER_QUERY_H_
